+++ src/cpu-template-helper/src/main.rs
@@ -58,7 +58,7 @@ enum TemplateOperation {
     Dump {
         /// Path of firecracker config file.
         #[arg(short, long, value_name = "PATH")]
-        config: PathBuf,
+        config: Option<PathBuf>,
         /// Path of output file.
         #[arg(short, long, value_name = "PATH", default_value = "cpu_config.json")]
         output: PathBuf,
@@ -76,7 +76,7 @@ enum TemplateOperation {
     Verify {
         /// Path of firecracker config file specifying the target CPU template.
         #[arg(short, long, value_name = "PATH")]
-        config: PathBuf,
+        config: Option<PathBuf>,
     },
 }
 
@@ -86,7 +86,7 @@ enum FingerprintOperation {
     Dump {
         /// Path of firecracker config file.
         #[arg(short, long, value_name = "PATH")]
-        config: PathBuf,
+        config: Option<PathBuf>,
         /// Path of output file.
         #[arg(short, long, value_name = "PATH", default_value = "fingerprint.json")]
         output: PathBuf,
@@ -115,8 +115,8 @@ fn run(cli: Cli) -> Result<(), HelperError> {
     match cli.command {
         Command::Template(op) => match op {
             TemplateOperation::Dump { config, output } => {
-                let config = read_to_string(config)?;
-                let (vmm, _) = utils::build_microvm_from_config(&config)?;
+                let config = config.map(read_to_string).transpose()?;
+                let (vmm, _) = utils::build_microvm_from_config(config)?;
 
                 let cpu_config = template::dump::dump(vmm)?;
 
@@ -140,8 +140,8 @@ fn run(cli: Cli) -> Result<(), HelperError> {
                 }
             }
             TemplateOperation::Verify { config } => {
-                let config = read_to_string(config)?;
-                let (vmm, vm_resources) = utils::build_microvm_from_config(&config)?;
+                let config = config.map(read_to_string).transpose()?;
+                let (vmm, vm_resources) = utils::build_microvm_from_config(config)?;
 
                 let cpu_template = vm_resources
                     .vm_config
@@ -155,8 +155,8 @@ fn run(cli: Cli) -> Result<(), HelperError> {
         },
         Command::Fingerprint(op) => match op {
             FingerprintOperation::Dump { config, output } => {
-                let config = read_to_string(config)?;
-                let (vmm, _) = utils::build_microvm_from_config(&config)?;
+                let config = config.map(read_to_string).transpose()?;
+                let (vmm, _) = utils::build_microvm_from_config(config)?;
 
                 let fingerprint = fingerprint::dump::dump(vmm)?;
 
@@ -195,69 +195,9 @@ fn main() -> Result<(), HelperError> {
 mod tests {
     use std::io::Write;
 
-    use vmm::utilities::mock_resources::kernel_image_path;
     use vmm_sys_util::tempfile::TempFile;
 
     use super::*;
-
-    pub fn generate_config(kernel_image_path: &str, rootfs_path: &str) -> String {
-        format!(
-            r#"{{
-                "boot-source": {{
-                    "kernel_image_path": "{}"
-                }},
-                "drives": [
-                    {{
-                        "drive_id": "rootfs",
-                        "path_on_host": "{}",
-                        "is_root_device": true,
-                        "is_read_only": false
-                    }}
-                ]
-            }}"#,
-            kernel_image_path, rootfs_path,
-        )
-    }
-
-    pub fn generate_config_with_template(
-        kernel_image_path: &str,
-        rootfs_path: &str,
-        cpu_template_path: &str,
-    ) -> String {
-        format!(
-            r#"{{
-                "boot-source": {{
-                    "kernel_image_path": "{}"
-                }},
-                "drives": [
-                    {{
-                        "drive_id": "rootfs",
-                        "path_on_host": "{}",
-                        "is_root_device": true,
-                        "is_read_only": false
-                    }}
-                ],
-                "cpu-config": "{}"
-            }}"#,
-            kernel_image_path, rootfs_path, cpu_template_path,
-        )
-    }
-
-    fn generate_config_file(
-        kernel_image_path: &str,
-        rootfs_path: &str,
-        cpu_template_path: Option<&str>,
-    ) -> TempFile {
-        let config = match cpu_template_path {
-            Some(cpu_template_path) => {
-                generate_config_with_template(kernel_image_path, rootfs_path, cpu_template_path)
-            }
-            None => generate_config(kernel_image_path, rootfs_path),
-        };
-        let config_file = TempFile::new().unwrap();
-        config_file.as_file().write_all(config.as_bytes()).unwrap();
-        config_file
-    }
 
     // Sample modifiers for x86_64 that should work correctly as a CPU template and a guest CPU
     // config.
@@ -334,21 +274,11 @@ mod tests {
 
     #[test]
     fn test_template_dump_command() {
-        let kernel_image_path = kernel_image_path(None);
-        let rootfs_file = TempFile::new().unwrap();
-        let config_file = generate_config_file(
-            &kernel_image_path,
-            rootfs_file.as_path().to_str().unwrap(),
-            None,
-        );
         let output_file = TempFile::new().unwrap();
-
         let args = vec![
             "cpu-template-helper",
             "template",
             "dump",
-            "--config",
-            config_file.as_path().to_str().unwrap(),
             "--output",
             output_file.as_path().to_str().unwrap(),
         ];
@@ -374,22 +304,7 @@ mod tests {
 
     #[test]
     fn test_template_verify_command() {
-        let kernel_image_path = kernel_image_path(None);
-        let rootfs_file = TempFile::new().unwrap();
-        let template_file = generate_sample_template();
-        let config_file = generate_config_file(
-            &kernel_image_path,
-            rootfs_file.as_path().to_str().unwrap(),
-            Some(template_file.as_path().to_str().unwrap()),
-        );
-
-        let args = vec![
-            "cpu-template-helper",
-            "template",
-            "verify",
-            "--config",
-            config_file.as_path().to_str().unwrap(),
-        ];
+        let args = vec!["cpu-template-helper", "template", "verify"];
         let cli = Cli::parse_from(args);
 
         run(cli).unwrap();
@@ -397,21 +312,11 @@ mod tests {
 
     #[test]
     fn test_fingerprint_dump_command() {
-        let kernel_image_path = kernel_image_path(None);
-        let rootfs_file = TempFile::new().unwrap();
-        let config_file = generate_config_file(
-            &kernel_image_path,
-            rootfs_file.as_path().to_str().unwrap(),
-            None,
-        );
         let output_file = TempFile::new().unwrap();
-
         let args = vec![
             "cpu-template-helper",
             "fingerprint",
             "dump",
-            "--config",
-            config_file.as_path().to_str().unwrap(),
             "--output",
             output_file.as_path().to_str().unwrap(),
         ];